+++ tests/dummy/tests/Feature/Avatars/CreateTest.php
@@ -47,9 +47,10 @@ class CreateTest extends TestCase
 
         $response = $this
             ->jsonApi()
+            ->includePaths('user')
             ->contentType($contentType)
-            ->content(['avatar' => $file])
-            ->post('/api/v1/avatars?include=user');
+            ->withPayload(['avatar' => $file])
+            ->post('/api/v1/avatars');
 
         $id = $response
             ->assertCreatedWithServerId(url('/api/v1/avatars'), $expected)